requirements.txt
# pip install  -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/


fastapi
uvicorn
psycopg[binary]
PyYAML
pytest
python-dotenv
httpx
